UserAuthAPI/requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==1.9
django-rest-authemail==1.4.0
djangorestframework==3.3.3
idna==2.7
requests==2.19.1
urllib3==1.23
